scripts/requirements.txt
alabaster==0.7.12
amqp==5.0.6
aniso8601==9.0.1
argh==0.26.2
astroid==2.15.4
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blueprint==3.4.2
boto3==1.16.21
celery==5.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.5
docutils==0.19
exceptiongroup==1.1.1
expiringdict==1.2.1
flake8==3.8.4
flake8-polyfill==1.0.2
h11==0.14.0
html2text==2020.1.16
humanize==3.9.0
idna==2.10
imagesize==1.2.0
iniconfig==2.0.0
isort==5.6.4
itsdangerous==1.1.0
jenkinsapi==0.3.11
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.2.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.33
numpy==1.21.2
outcome==1.2.0
packaging==20.4
pathtools==0.1.2
pdfkit==0.6.1
pep257==0.7.0
pep8-naming==0.10.0
pkcs7==0.1.2
pluggy==1.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.20.3
py==1.11.0
py4j==0.10.9
pycodestyle==2.6.0
pycryptodome==3.9.9
pycurl==7.43.0.6
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
# pyspark==3.1.2
pytest==7.1.2
python-dotenv==1.0.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
requirements-detector==0.7
s3transfer==0.3.3
scipy==1.7.3
selenium==4.4.1
six==1.15.0
sniffio==1.3.0
snowballstemmer==2.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
ua-parser==0.10.0
unicodecsv==0.14.1
urllib3==1.26.2
uwsgitop==0.11
vine==5.0.0
watchdog==0.10.3
wcwidth==0.2.5
webdriver-manager==3.8.6
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==1.3.7
stripe==5.4.0
xmltodict==0.13.0
openpyxl==3.1.2
lxml==4.9.2
pandas==2.0.1
